update 28.11.2023
- started working on the role model of content management in the application

+++ src/main/java/lab/fcpsr/suprime/templates/MinioResponse.java
@@ -10,6 +10,6 @@ public class MinioResponse {
     private String originalFileName;
     private String uid;
     private String type;
-    private int fileSize;
+    private float fileSize;
     private ObjectWriteResponse response;
 }
