+++ src/main/java/lab/fcpsr/suprime/utils/CustomFileUtil.java
@@ -3,14 +3,14 @@ package lab.fcpsr.suprime.utils;
 import java.util.Optional;
 
 public class CustomFileUtil {
-    private static final int ONE_MB = 1048576;
+    private static final float ONE_MB = 1048576.0f;
     public static Optional<String> getExtension(String filename){
         return Optional.ofNullable(filename)
                 .filter(f -> f.contains("."))
                 .map(f -> f.substring(filename.lastIndexOf(".") + 1));
     }
 
-    public static int getMegaBytes(int bitsAmount){
+    public static float getMegaBytes(int bitsAmount){
         return bitsAmount / ONE_MB;
     }
 }
